+++ CHANGELOG.md
@@ -6,9 +6,12 @@
 
 - **Adjusted _"Seizure"_ screen wipe's speed** to match _"Melt"_ and _"Fade"_
 - **Automap position isn't reset to player position when opening it with Follow Mode off**
+- **FOV is now changed gradually** in most cases
+- **Reduced mouse sensitivity when zoomed in**
 - **Renamed _"Imitate player's breathing"_ to _"Subtle Idle Bobbing/Breathing"_**
 
 ## Bug Fixes
 
 - **Made _"Seizure"_ screen wipe's speed consistent** in low and high resolution
 - **Fixed some cheat activation messages**
+- **Zoom FOV is now unaffected by changes to normal FOV, and vice versa**